requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.6
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
docker==6.1.3
docstring-parser==0.15
executing==2.0.1
fastjsonschema==2.18.1
FLAML==2.1.1
fonttools==4.44.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.344
langchain-core==0.0.8
langsmith==0.0.68
llama-index==0.8.69.post2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
openai==1.3.7
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prettytable==3.9.0
prometheus-client==0.18.0
prompt-toolkit==3.0.36
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyautogen==0.2.27
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pymemgpt==0.2.5
pyparsing==3.1.1
pypdf==3.17.1
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
questionary==2.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.12.0
scipy==1.11.3
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
splunk-sdk==1.7.4
SQLAlchemy==2.0.23
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
tiktoken==0.5.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.9.2
yfinance==0.2.31
